+++ simple_value_object/exceptions.py
@@ -1,23 +1,23 @@
-class NotDeclaredArgsException(Exception):
-    def __init__(self, *args, **kwargs):
-        super(NotDeclaredArgsException, self).__init__(
-            'No arguments declared in __init__'
-        )
+class SimpleValueObjectException(Exception):
+    pass
 
 
-class CannotBeChangeException(Exception):
-    def __init__(self, *args, **kwargs):
-        super(CannotBeChangeException, self).__init__(
+class ConstructorWithoutArguments(SimpleValueObjectException):
+    def __init__(self):
+        super().__init__('No arguments declared in __init__')
+
+
+class CannotBeChanged(SimpleValueObjectException):
+    def __init__(self):
+        super().__init__(
             'You cannot change values from a Value Object, create a new one'
         )
 
 
-class ViolatedInvariantException(Exception):
+class InvariantViolation(SimpleValueObjectException):
     pass
 
 
-class InvariantReturnValueException(Exception):
-    def __init__(self, *args, **kwargs):
-        super(InvariantReturnValueException, self).__init__(
-            'Invariants must return a boolean value'
-        )
+class InvariantMustReturnBool(SimpleValueObjectException):
+    def __init__(self):
+        super().__init__('Invariants must return a boolean value')